Add BootloaderConnection.erase_flash to API
erase_flash is a new high-level interface which calls _erase_flash in
the background. It erases the entire program memory area by default,
but can be told to only erase a smaller part.

By default, erase_flash checks whether the program memory area
contains a program before and after erasing. If no program is detected
before erasing, the erase is skipped. If a program is detected after
erasing, an exception is raised.