Add get_response function and deprecate Packet.from_serial
Turns out error responses are not guaranteed to be the same length as
a packet of the expected response type. Relying on the command code to
predict the length of the response is therefore not viable. get_packet
instead reads the first byte and uses it to determine the type of the
response.