Split reading the response into several steps
First get the command echo; the bootloader must reply with at least
that many bytes.

If the command was READ_VERSION, read the rest of the data and return
it.

Otherwise, read one more byte to get the success flag. If it is
anything but SUCCESS, raise appropriate exception. Otherwise, read the
rest of the data and return it.

Also, print way more raw data in debug log.